Type conversions

+++ ai_lab2/src/ai_lab2/Item.java
@@ -1,32 +1,32 @@
 package ai_lab2;
 
 public class Item {
-	private int value;
-	private int weight;
+	private double value;
+	private double weight;
 	
-	public Item(int value, int weight) {
+	public Item(double value, double weight) {
 		this.value = value;
 		this.weight = weight;
 	}
 	
 	//another solution ---> give random values to weight resp. value within given interval?
-	public void setValue(int value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
 	
-	public void setWeight(int weight) {
+	public void setWeight(double weight) {
 		this.weight = weight;
 	}
 	
-	public int getValue() {
+	public double getValue() {
 		return value;
 	}
 	
-	public int getWeight() {
+	public double getWeight() {
 		return weight;
 	}
 	
-	public int calcBenefit() {
+	public double calcBenefit() {
 		return (value / weight);
 	}
